pyproject.toml
[project]
name = "trading-simulator"
version = "0.0.1"
description = "A project description"
authors = [
    {name = "Michał Rutkowski", email = "homeend@wp.pl"},
]
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "django==5.1.4",
    "django-environ==0.11.2",
    "django-structlog==9.0.0",
    "honcho>=2.0.0",
    "structlog>=24.4.0",
    "whitenoise==6.8.2",
    "more-itertools==10.6.0",
    "drf-yasg>=1.21.10",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "arrow>=1.3.0",
    "prometheus-client>=0.22.0",
    "django-prometheus>=2.3.1",
    "django-business-metrics>=1.0.1",
    "django-health-check>=3.18.3",
    "psutil>=7.0.0",
    "pydantic>=2.11.4",
    "django-extensions>=4.1",
    "django-probes>=1.7.0",
    "django-celery-beat>=2.8.1",
    "redis>=6.1.0",
    "uvicorn[standard]>=0.34.2",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "assertpy>=1.1",

    "pytest>=8.3.5",
    "pytest-django>=4.11.1",


    "time-machine>=2.16.0",

]


[tool.black]
line-length = 120
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.nox
  | \.tox
  | \.venv
  | \._build
  | buck-out
  | build
  | dist
  # Add this line to exclude all migration directories
  | migrations
  # Or, if you want to be more specific to only Python files within migrations:
  # | .*/migrations/.*\.py
)/
'''

